Refactor database functionality and streamline Lifestream class for improved connection handling
